Stop listening to `push` event
The `push` event happens a lot and listening to it might increase the risk of getting "API rate limit exceeded" responses when calling the GitHub REST API.
Instead, we listen to `pull_request.closed` event. It should be enough because, in sane projects with branch protections set up, `master` should only move forward by merging pull requests on it.